backend/requirements.txt
click==8.1.8
dnspython==2.3.0
h11==0.14.0
importlib-metadata==6.7.0
numpy==1.21.6
pandas==1.3.5
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.2
six==1.17.0
typing-extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
fastapi
python-dateutil